feat: Support IPv6 for backend server
It has been a while since IPv6 launch day 6.6.2012, so it the NodeRed backend
should support IPv6 too. Nearly all systems default to IPv6 and also names like
localhost are resolved to their IPv6 address by default. This means if a user
enters localhost in NodeJS for the server config it will fail. Fix this by
listen to IPv4 and IPv6.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>